Start typeorm splitting
