+++ packages/database-typeorm/src/typeorm.ts
@@ -1,24 +1,16 @@
-import { ConnectionInformations, CreateUser, DatabaseInterface } from '@accounts/types';
+import { CreateUser, DatabaseInterface } from '@accounts/types';
 import { Repository, getRepository } from 'typeorm';
 import { User } from './entity/User';
 import { UserEmail } from './entity/UserEmail';
 import { UserService } from './entity/UserService';
-import { UserSession } from './entity/UserSession';
 import { AccountsTypeormOptions } from './types';
-
-const defaultOptions = {
-  userEntity: User,
-  userEmailEntity: UserEmail,
-  userServiceEntity: UserService,
-  userSessionEntity: UserSession,
-};
+import { defaultOptions } from './options';
 
 export class AccountsTypeorm implements DatabaseInterface {
   private options: AccountsTypeormOptions & typeof defaultOptions;
   private userRepository: Repository<User> = null as any;
   private emailRepository: Repository<UserEmail> = null as any;
   private serviceRepository: Repository<UserService> = null as any;
-  private sessionRepository: Repository<UserSession> = null as any;
 
   constructor(options?: AccountsTypeormOptions) {
     this.options = { ...defaultOptions, ...options };
@@ -29,7 +21,6 @@ export class AccountsTypeorm implements DatabaseInterface {
       userEntity,
       userEmailEntity,
       userServiceEntity,
-      userSessionEntity,
     } = this.options;
 
     const setRepositories = () => {
@@ -37,12 +28,10 @@ export class AccountsTypeorm implements DatabaseInterface {
         this.userRepository = connection.getRepository(userEntity);
         this.emailRepository = connection.getRepository(userEmailEntity);
         this.serviceRepository = connection.getRepository(userServiceEntity);
-        this.sessionRepository = connection.getRepository(userSessionEntity);
       } else {
         this.userRepository = getRepository(userEntity, connectionName);
         this.emailRepository = getRepository(userEmailEntity, connectionName);
         this.serviceRepository = getRepository(userServiceEntity, connectionName);
-        this.sessionRepository = getRepository(userSessionEntity, connectionName);
       }
     };
 
@@ -89,38 +78,6 @@ export class AccountsTypeorm implements DatabaseInterface {
       return null;
     }
     return user;
-  }
-
-  public async findUserByResetPasswordToken(token: string): Promise<User | null> {
-    const service = await this.serviceRepository.findOne({
-      where: {
-        name: 'password.reset',
-        token,
-      },
-      cache: this.options.cache,
-    });
-
-    if (service) {
-      return this.findUserById(service.userId);
-    }
-
-    return null;
-  }
-
-  public async findUserByEmailVerificationToken(token: string): Promise<User | null> {
-    const service = await this.serviceRepository.findOne({
-      where: {
-        name: 'email.verificationTokens',
-        token,
-      },
-      cache: this.options.cache,
-    });
-
-    if (service) {
-      return this.findUserById(service.userId);
-    }
-
-    return null;
   }
 
   public async createUser(createUser: CreateUser): Promise<string> {
@@ -235,196 +192,11 @@ export class AccountsTypeorm implements DatabaseInterface {
     }
   }
 
-  public async findPasswordHash(userId: string): Promise<string | null> {
-    const service = await this.getService(userId, 'password');
-    if (service) {
-      return service.options.bcrypt;
-    }
-    return null;
-  }
-
-  public async setPassword(userId: string, newPassword: string): Promise<void> {
-    const user = await this.findUserById(userId);
-    if (user) {
-      await this.setService(userId, 'password', { bcrypt: newPassword });
-      await this.userRepository.update({ id: user.id }, {});
-      return;
-    }
-    throw new Error('User not found');
-  }
-
-  public async addResetPasswordToken(
-    userId: string,
-    email: string,
-    token: string,
-    reason: string
-  ): Promise<void> {
-    await this.setService(
-      userId,
-      'password.reset',
-      {
-        address: email.toLocaleLowerCase(),
-        when: new Date().toJSON(),
-        reason,
-      },
-      token
-    );
-  }
-
-  public async setResetPassword(
-    userId: string,
-    email: string,
-    newPassword: string,
-    token?: string
-  ): Promise<void> {
-    await this.setPassword(userId, newPassword);
-    await this.unsetService(userId, 'password.reset');
-  }
-
-  public async addEmail(userId: string, newEmail: string, verified: boolean): Promise<void> {
-    const user = await this.findUserById(userId);
-    if (user) {
-      const userEmail = new this.options.userEmailEntity();
-      userEmail.user = user;
-      userEmail.address = newEmail.toLocaleLowerCase();
-      userEmail.verified = verified;
-      await this.emailRepository.save(userEmail);
-      await this.userRepository.update({ id: user.id }, {});
-      return;
-    }
-    throw new Error('User not found');
-  }
-
-  public async removeEmail(userId: string, email: string): Promise<void> {
-    const user = await this.findUserById(userId);
-    if (user) {
-      const userEmail = user.emails.find(s => s.address === email.toLocaleLowerCase());
-      if (!userEmail) {
-        throw new Error('Email not found');
-      }
-      await this.emailRepository.remove(userEmail);
-      await this.userRepository.update({ id: user.id }, {});
-      return;
-    }
-    throw new Error('User not found');
-  }
-
-  public async verifyEmail(userId: string, email: string): Promise<void> {
-    const user = await this.findUserById(userId);
-    if (user) {
-      const userEmail = user.emails.find(s => s.address === email.toLocaleLowerCase());
-
-      if (!userEmail) {
-        throw new Error('Email not found');
-      }
-      userEmail.verified = true;
-      await this.emailRepository.save(userEmail);
-      await this.unsetService(userId, 'email.verificationTokens');
-      await this.userRepository.update({ id: user.id }, {});
-      return;
-    }
-    throw new Error('User not found');
-  }
-
-  public async addEmailVerificationToken(
-    userId: string,
-    email: string,
-    token: string
-  ): Promise<void> {
-    await this.setService(
-      userId,
-      'email.verificationTokens',
-      {
-        address: email.toLocaleLowerCase(),
-        when: new Date(),
-      },
-      token
-    );
-  }
-
   public async setUserDeactivated(userId: string, deactivated: boolean): Promise<void> {
     const user = await this.findUserById(userId);
     if (user) {
       user.deactivated = deactivated;
       await this.userRepository.save(user);
     }
-  }
-
-  public async findSessionById(sessionId: string): Promise<UserSession | null> {
-    try {
-      const session = await this.sessionRepository.findOne(sessionId, {
-        cache: this.options.cache,
-      });
-
-      if (session) {
-        return session;
-      }
-    } catch (err) {
-      // noop
-    }
-
-    return null;
-  }
-
-  public async findSessionByToken(token: string) {
-    const session = await this.sessionRepository.findOne(
-      { token },
-      {
-        cache: this.options.cache,
-      }
-    );
-    if (!session) {
-      return null;
-    }
-    return session;
-  }
-
-  public async createSession(
-    userId: string,
-    token: string,
-    connection: ConnectionInformations = {},
-    extra?: object
-  ) {
-    const user = await this.findUserById(userId);
-    const session = new this.options.userSessionEntity();
-    session.user = user!;
-    session.token = token;
-    session.userAgent = connection.userAgent;
-    session.ip = connection.ip;
-    if (extra) {
-      session.extra = extra;
-    }
-    session.valid = true;
-    await this.sessionRepository.save(session);
-
-    return session.id;
-  }
-
-  public async updateSession(sessionId: string, connection: ConnectionInformations): Promise<void> {
-    const session = await this.findSessionById(sessionId);
-    if (session) {
-      session.userAgent = connection.userAgent;
-      session.ip = connection.ip;
-      await this.sessionRepository.save(session);
-    }
-  }
-
-  public async invalidateSession(sessionId: string): Promise<void> {
-    const session = await this.findSessionById(sessionId);
-    if (session) {
-      session.valid = false;
-      await this.sessionRepository.save(session);
-    }
-  }
-
-  public async invalidateAllSessions(userId: string): Promise<void> {
-    await this.sessionRepository.update(
-      {
-        userId,
-      },
-      {
-        valid: false,
-      }
-    );
   }
 }
